Restructuring header files.
A large number of files have been modified in order to ensure that each header file includes any C++ Standard Library headers it references. This avoids breaking the compile when GTE headers are refactored. I also unified the commenting so that top-level comments occur after the pragma-once statement but before any header file includes. (All of this had already been done in GTL and ensures that the instantiation tool behaves correctly.) See Gte6UpdateHistory.pdf for the details of all the changes.

+++ GTE/Mathematics/NaturalCubicSpline.h
@@ -3,22 +3,23 @@
 // Distributed under the Boost Software License, Version 1.0.
 // https://www.boost.org/LICENSE_1_0.txt
 // https://www.geometrictools.com/License/Boost/LICENSE_1_0.txt
-// Version: 6.0.2022.06.07
+// Version: 6.0.2023.08.08
 
 #pragma once
-
-#include <Mathematics/Matrix.h>
-#include <Mathematics/Matrix3x3.h>
-#include <Mathematics/ParametricCurve.h>
-#include <array>
-#include <cstdint>
-#include <vector>
 
 // Documentation for natural splines is found in
 // https://www.geometrictools.com/Documentation/NaturalSplines.pdf
 // The number of points must be 2 or larger. The points[] and times[] arrays
 // must have the same number of elements. The times[] values must be strictly
 // increasing.
+
+#include <Mathematics/Matrix.h>
+#include <Mathematics/Matrix3x3.h>
+#include <Mathematics/ParametricCurve.h>
+#include <array>
+#include <cstddef>
+#include <cstdint>
+#include <vector>
 
 namespace gte
 {
@@ -255,7 +256,7 @@ namespace gte
                 T LUProd0 = r2 * sigma, LUProd1 = -sigmasqr;
                 T sign = -r1;
 
-                for (int32_t i = 1; i <= numPm3; ++i)
+                for (size_t i = 1; i <= static_cast<size_t>(numPm3); ++i)
                 {
                     Btrg -= sign * (LUProd0 * B[3 * i] + LUProd1 * B[3 * i + 1]);
                     sigma = mDelta[i] / mDelta[i + 1];
@@ -280,7 +281,7 @@ namespace gte
                 T LUProd0 = -r3 * sigma, LUProd1 = r2 * sigmasqr;
                 T sign = -r1;
 
-                for (int32_t i = 1; i <= numPm3; ++i)
+                for (size_t i = 1; i <= static_cast<size_t>(numPm3); ++i)
                 {
                     Btrg -= sign * (LUProd0 * B[3 * i] + LUProd1 * B[3 * i + 1]);
                     sigma = mDelta[i] / mDelta[i + 1];
